Massage incoming record so `json` is always a string
The backend behaves oddly here: if the field `json` is set to a string
that can be parsed as JSON, then that is done and fetching the record
returns a parsed object as the value of the field; if it's set to
anything else, that is returned as a string. Now in the former case,
we just serialise back to a string which we can edit.