Add instructure for help-texts in each editable field
Half a dozen texts have been added as proof of concept; the rest will
be ported over from the legacy harvester in due course.

This approach works well for `<TextField>`s and some other controls,
but not for `<TextArea>`s and some others. This seems to be because of
the unever implementation of the `endControl` prop across some but not
all of the stripes-components form components.