Temporarily disable attempts at OAI-PMH
These are doomed due to CORS, so there is no point in flooding the
console with worrying errors.

In the longer term, if we want to do this, we will probably need to
proxy via the FOLIO service, either using dedicated WSAPIs provided by
mod-harvester-admin or a general HTTP-proxy module.