Status is once more displayed in Job pane-title
This is done in a slightly worrying way, by consulting both
`record.currentStatus` (as before) and `record.status` (which is what
we get in the record). I suspect this has changed in the back-end, so
we may have to revisit this (or other uses of status) once we've
figured out the back-story.

Fixes UIHAADM-120.