Improve the situation with keeping all tabs on the page (#67)
We are now down to two scrollbars instead of three(!), and the Harvestables/Jobs/Failed Records tabs are now always on screen. But they can still be scrolled upwards and the FOLIO top-bar scrolled off the screen. This is not as good as the situation in the Agreements application, where things work as they ought.

I will return to this later and try to complete the work by eliminating the second scrollbar.

The present commit addresses but does not fully fix UIHAADM-99.