Failed-record errors from "message" instead of "label" (#71)
When a error structure has a `message` as well as a `label`, use the former rather than the latter. We still fall back to the label when there is no message.

Add tests for both cases.

These are in fact the first tests in the module, so now for the first time it's possible to run `yarn test`. A little bit of infrastructure was added to support the use of Jest.

Fixes UIHAADM-125.