Send full datetime values as part of query
When sending dates as part of a query, they are now extended to
complete datetimes. Start-of-range dates are extended to the beginning
of the specified day (`T00:00:00`) and end-of-range dates to the end
of the specified day (`T23:59:59`).

This means that the ends of date-ranges now include the whole of the
specified day rather than only its first second.

Fixes UIHAADM-126.