Fix CSV export not to use stripes-connect
Unfortunately, the original solution to including all the failed records in the CSV download, which used stripes-connect's `GET` mutator to fetch all the records in the result set,  didn't work out. It somehow left stripes-connect in strange states where it would try to display the page of records past the end of the result set, or claim that there were NaN records.

I spent a little while trying to persuade stripes-connect to get back into a sensible state after doing the export — thought I don't really understand why it was in an odd state to start with — but couldn't make it work either with the `reset` mutator or with another `GET` of the first page.

Rather than dive into the horrors of stripes-connect's Redux store, which is where its state lives, I made the pragmatic decision just not to use stripes-connect for this, and instead to fetch the records imperatively using `okapiKy`. This works out fine.